+++ src/APC524/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations  # MUST be first
 from . import solver, visualization
 from .solver.automaton import CellularAutomaton
 from .solver.kernels import MOORE_KERNEL
-from .solver.rules import CGOL_rules
+from .solver.rules import CGOL_init, CGOL_rules
 from .solver.utils import convolve_neighbours_2D
 from .visualization.visualization import animate_automaton
 
@@ -13,6 +13,7 @@ __all__ = [
     "CellularAutomaton",
     "MOORE_KERNEL",
     "CGOL_rules",
+    "CGOL_init",
     "convolve_neighbours_2D",
     "animate_automaton",
 ]